requirements.txt
-i https://pypi.org/simple
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
click==7.0
flask==1.0.3
itsdangerous==1.1.0
jinja2==2.10.1
kombu==4.6.1
markupsafe==1.1.1
python-dotenv==0.10.3
pytz==2019.1
redis==3.2.1
vine==1.3.0
werkzeug==0.15.4
